stream-processing/alpaca-crypto-client/pyproject.toml
[tool.poetry]
name = "alpaca-crypto-client"
version = "0.1.0"
description = ""
authors = ["andersonvc <ryan@anderson.vc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
alpaca-py = "^0.21.0"
avro = "^1.11.3"
confluent-kafka = "^2.3.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# add script
[tool.poetry.scripts]
main = "alpaca_crypto_client.main:start"
